Change integer in biginteger

+++ database/migrations/2023_02_20_101944_create_movies_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
 
             $table->string('name', 64);
             $table->year('year');
-            $table->integer('cashOut')->unsigned()->nullable();
+            $table->bigInteger('cashOut')->unsigned()->nullable();
 
             $table->timestamps();
         });
